CryptoProject/app/encryption.py
from Crypto.Cipher import AES # type: ignore
from Crypto.Util.Padding import pad, unpad # type: ignore
import os

def generate_key():
    return os.urandom(32)  # AES-256 key

def encrypt_message(key, plaintext):
    cipher = AES.new(key, AES.MODE_CBC)
    iv = cipher.iv # type: ignore
    ciphertext = cipher.encrypt(pad(plaintext.encode(), AES.block_size))
    return iv + ciphertext # type: ignore

def decrypt_message(key, ciphertext):
    iv = ciphertext[:16]
    actual_ciphertext = ciphertext[16:]
    cipher = AES.new(key, AES.MODE_CBC, iv)
    plaintext = unpad(cipher.decrypt(actual_ciphertext), AES.block_size)
    return plaintext.decode()
