CryptoProject/requirements.txt
Flask==2.0.1
pycryptodome==3.10.1
twilio==6.59.0
